Wenn wir jetzt *nicht* mehr Reflection machen, *dann* wird die Factory auf einmal wieder interessant. Dann ist nämlich die *Factory* das, was man austauscht, und die selber weiß, was es für SpatialRules es gibt, und die man *selber* ggf. per Reflection austauschbar machen könnte, wenn's sein muss, *statt* der einzelnen Rules. Die kann dadurch dann auch viel mehr: Nämlich selber entscheiden, ob evtl. mehrere Länder die gleiche Regel haben sollen, oder ob sie selber einfach ad-hoc-Implementierungen des Interfaces erzeugt, etc.
