+++ core/src/test/java/com/graphhopper/routing/util/DataFlagEncoderTest.java
@@ -2,15 +2,9 @@ package com.graphhopper.routing.util;
 
 import java.util.*;
 
-import com.graphhopper.json.geo.GeoJsonPolygon;
-import com.graphhopper.json.geo.Geometry;
-import com.graphhopper.json.geo.JsonFeature;
-import com.graphhopper.json.geo.JsonFeatureCollection;
 import com.graphhopper.routing.AbstractRoutingAlgorithmTester;
 import com.graphhopper.util.PMap;
 import com.graphhopper.routing.util.spatialrules.*;
-import com.graphhopper.routing.util.spatialrules.countries.GermanySpatialRule;
-import com.graphhopper.util.shapes.BBox;
 import com.graphhopper.util.shapes.GHPoint;
 import org.junit.Test;
 
@@ -318,7 +312,7 @@ public class DataFlagEncoderTest {
 
     @Test
     public void testSpatialId() {
-        final GermanySpatialRule germany = new GermanySpatialRule();
+        final SpatialRule germany = new GermanySpatialRule();
         germany.setBorders(Collections.singletonList(new Polygon(new double[]{0, 0, 1, 1}, new double[]{0, 1, 1, 0})));
 
         SpatialRuleLookup index = new SpatialRuleLookup() {
@@ -402,5 +396,49 @@ public class DataFlagEncoderTest {
 
         assertEquals(5, encoder.getMaxspeed(e3, -1, false), .1);
         assertEquals(-1, encoder.getMaxspeed(e4, -1, false), .1);
+    }
+
+    /**
+     * Defines the default rules for German roads
+     *
+     * @author Robin Boldt
+     */
+    private static class GermanySpatialRule extends DefaultSpatialRule {
+
+        /**
+         * Germany contains roads with no speed limit. For these roads, this method will return Integer.MAX_VALUE.
+         * Your implementation should be able to handle these cases.
+         */
+        @Override
+        public double getMaxSpeed(String highwayTag, double _default) {
+            // As defined in: https://wiki.openstreetmap.org/wiki/OSM_tags_for_routing/Maxspeed#Motorcar
+            switch (highwayTag) {
+                case "motorway":
+                    return Integer.MAX_VALUE;
+                case "trunk":
+                    return Integer.MAX_VALUE;
+                case "residential":
+                    return 100;
+                case "living_street":
+                    return 4;
+                default:
+                    return super.getMaxSpeed(highwayTag, _default);
+            }
+        }
+
+        @Override
+        public AccessValue getAccessValue(String highwayTag, TransportationMode transportationMode, AccessValue _default) {
+            if (transportationMode == TransportationMode.MOTOR_VEHICLE) {
+                if (highwayTag.equals("track"))
+                    return AccessValue.EVENTUALLY_ACCESSIBLE;
+            }
+
+            return super.getAccessValue(highwayTag, transportationMode, _default);
+        }
+
+        @Override
+        public String getId() {
+            return "DEU";
+        }
     }
 }
